requirements.txt
Flask==2.3.2
setuptools==58.0.0
wheel
numpy==1.24.4
pandas==2.0.3
pymysql==1.0.2
python-dotenv==1.0.0